remove graduation-status-container
